+++ controlgallery/AtomUIGallery/ShowCases/General/Window/ShowCaseControls/AddOnDemoWindow.axaml.cs
@@ -1,0 +1,9 @@
+namespace AtomUIGallery.ShowCases.Window;
+
+internal partial class AddOnDemoWindow : AtomUI.Desktop.Controls.Window
+{
+    public AddOnDemoWindow()
+    {
+        InitializeComponent();
+    }
+}