+++ controlgallery/AtomUIGallery/ShowCases/General/Window/ShowCaseControls/BasicDemoWindow.axaml.cs
@@ -1,0 +1,9 @@
+namespace AtomUIGallery.ShowCases.Window;
+
+internal partial class BasicDemoWindow : AtomUI.Desktop.Controls.Window
+{
+    public BasicDemoWindow()
+    {
+        InitializeComponent();
+    }
+}